Cleanup, formatting, etc.
