+++ src/repl/config.rs
@@ -19,7 +19,6 @@ pub(crate) static DEFAULT_CONFIG: &str = include_str!(concat!(
 
 pub(crate) struct ReplConfig {
     pub config: Config,
-    pub config_dir: PathBuf,
     pub cache_dir: PathBuf,
     pub history_path: PathBuf,
 }
@@ -28,8 +27,8 @@ pub(crate) fn repl_config() -> ReplConfig {
     let dirs = ProjectDirs::from("org", "partiql", "partiql-cli").expect("project directories");
     let config_dir = dirs.config_dir();
     let cache_dir = dirs.cache_dir();
-    std::fs::create_dir_all(&config_dir).expect("create config dir");
-    std::fs::create_dir_all(&cache_dir).expect("create cache dir");
+    std::fs::create_dir_all(config_dir).expect("create config dir");
+    std::fs::create_dir_all(cache_dir).expect("create cache dir");
 
     let mut conf = config_dir.to_path_buf();
     conf.push("partiql-cli.toml");
@@ -37,7 +36,8 @@ pub(crate) fn repl_config() -> ReplConfig {
     // If the config file does not exist, create it and write the default config into it.
     //    create_new returns `Ok` if it creates and `Err` if the file already exists
     if let Ok(mut f) = OpenOptions::new().write(true).create_new(true).open(&conf) {
-        f.write_all(DEFAULT_CONFIG.as_bytes());
+        f.write_all(DEFAULT_CONFIG.as_bytes())
+            .expect("write default config");
     }
 
     let config = Config::builder()
@@ -62,14 +62,13 @@ pub(crate) fn repl_config() -> ReplConfig {
         .create_new(true)
         .open(&history_path)
     {
-        if let Ok(mut lf) = OpenOptions::new().read(true).open(&legacy_history_path) {
+        if let Ok(mut lf) = OpenOptions::new().read(true).open(legacy_history_path) {
             std::io::copy(&mut lf, &mut f).expect("copy legacy history");
         }
     }
 
     ReplConfig {
         config,
-        config_dir: config_dir.to_path_buf(),
         cache_dir: cache_dir.to_path_buf(),
         history_path,
     }
